+++ taipy/gui_core/_adapters.py
@@ -43,6 +43,8 @@ from taipy.gui._warnings import _warn
 from taipy.gui.gui import _DoNotUpdate
 from taipy.gui.utils import _is_boolean, _is_true, _TaipyBase
 
+from .filters import DataNodeFilter, ScenarioFilter, _Filter
+
 
 # prevent gui from trying to push scenario instances to the front-end
 class _GuiCoreDoNotUpdate(_DoNotUpdate):
@@ -353,102 +355,24 @@ def _get_entity_property(col: str, *types: t.Type):
         # we compare only strings
         if isinstance(entity, types):
             if isinstance(entity, Cycle):
-                lcol = "creation_date"
-                lfn = None
+                the_col = "creation_date"
+                the_fn = None
             else:
-                lcol = col
-                lfn = col_fn
+                the_col = col
+                the_fn = col_fn
             try:
-                val = attrgetter(lfn or lcol)(entity)
-                if lfn:
+                val = attrgetter(the_fn or the_col)(entity)
+                if the_fn:
                     val = val()
             except AttributeError as e:
                 if _is_debugging():
-                    _warn("Attribute", e)
+                    _warn(f"sort_key({entity.id}):", e)
                 val = ""
         else:
             val = ""
         return val.isoformat() if isinstance(val, (datetime, date)) else str(val)
 
     return sort_key
-
-
-@dataclass
-class _Filter(_DoNotUpdate):
-    label: str
-    property_type: t.Optional[t.Type]
-
-    def get_property(self):
-        return self.label
-
-    def get_type(self):
-        if self.property_type is bool:
-            return "boolean"
-        elif self.property_type is int or self.property_type is float:
-            return "number"
-        elif self.property_type is datetime or self.property_type is date:
-            return "date"
-        elif self.property_type is str:
-            return "str"
-        return "any"
-
-
-@dataclass
-class ScenarioFilter(_Filter):
-    """
-    used to describe a filter on a scenario property
-    """
-    property_id: str
-
-    def get_property(self):
-        return self.property_id
-
-
-@dataclass
-class DataNodeScenarioFilter(_Filter):
-    """
-    used to describe a filter on a scenario datanode's property
-    """
-    datanode_config_id: str
-    property_id: str
-
-    def get_property(self):
-        return f"{self.datanode_config_id}.{self.property_id}"
-
-
-_CUSTOM_PREFIX = "fn:"
-
-
-@dataclass
-class CustomScenarioFilter(_Filter):
-    """
-    used to describe a custom scenario filter ie based on a user defined function
-    """
-    filter_function: t.Callable[[Scenario], t.Any]
-
-    def __post_init__(self):
-        if self.filter_function.__name__ == "<lambda>":
-            raise TypeError("CustomScenarioFilter does not support lambda functions.")
-        mod = self.filter_function.__module__
-        self.module = mod if isinstance(mod, str) else mod.__name__
-
-    def get_property(self):
-        return f"{_CUSTOM_PREFIX}{self.module}:{self.filter_function.__name__}"
-
-    @staticmethod
-    def _get_custom(col: str) -> t.Optional[t.List[str]]:
-        return col[len(_CUSTOM_PREFIX) :].split(":") if col.startswith(_CUSTOM_PREFIX) else None
-
-
-@dataclass
-class DataNodeFilter(_Filter):
-    """
-    used to describe a filter on a datanode property
-    """
-    property_id: str
-
-    def get_property(self):
-        return self.property_id
 
 
 @dataclass(frozen=True)
